requirements.txt
requests==1.2.3
pyyaml==3.10